+++ src/SchemalessAttributesTrait.php
@@ -3,6 +3,7 @@
 namespace Spatie\SchemalessAttributes;
 
 use Illuminate\Support\Collection;
+use Spatie\SchemalessAttributes\Casts\SchemalessAttributes as SchemalessAttributesCast;
 
 /**
  * @property array $schemalessAttributes
@@ -14,13 +15,15 @@ trait SchemalessAttributesTrait
     public function initializeSchemalessAttributesTrait()
     {
         foreach ($this->getSchemalessAttributes() as $attribute) {
-            $this->casts[$attribute] = 'array';
+            $this->casts[$attribute] = SchemalessAttributesCast::class;
         }
     }
 
     public function getSchemalessAttributes(): array
     {
-        return isset($this->schemalessAttributes) ? $this->schemalessAttributes : [];
+        return isset($this->schemalessAttributes)
+            ? $this->schemalessAttributes
+            : [];
     }
 
     /**
@@ -29,7 +32,7 @@ trait SchemalessAttributesTrait
      */
     public function __get($key)
     {
-        if (in_array($key, $this->getSchemalessAttributes())) {
+        if (in_array($key, $this->getSchemalessAttributes(), true)) {
             return SchemalessAttributes::createForModel($this, $key);
         }