+++ tests/TestCase.php
@@ -23,11 +23,25 @@ abstract class TestCase extends Orchestra
         ];
     }
 
+    /*
+     * uses different table names for each test class to support
+     * running tests in parallel.
+     */
     protected function setUpDatabase()
     {
-        Schema::dropIfExists('test_models');
+        Schema::dropIfExists("test_models");
 
-        Schema::create('test_models', function (Blueprint $table) {
+        Schema::create("test_models", function (Blueprint $table) {
+            $table->increments('id');
+            $table->schemalessAttributes();
+        });
+        
+        $parts = explode('\\', static::class);
+        $class = array_pop($parts);
+        
+        Schema::dropIfExists("test_models_{$class}");
+
+        Schema::create("test_models_{$class}", function (Blueprint $table) {
             $table->increments('id');
             $table->schemalessAttributes();
         });